+++ learning/stdp/stdp_test.go
@@ -181,7 +181,7 @@ func TestWindow(t *testing.T) {
 	tests := []struct {
 		name string
 		dt   uint32
-		amp  int16
+		amp  int32
 		tau  uint32
 		want bool
 	}{
